+++ animation_nodes/ui/node_menu.py
@@ -426,6 +426,8 @@ class MeshDataMenu(bpy.types.Menu):
     def draw(self, context):
         layout = self.layout
         insertNode(layout, "an_VertexGroupInputNode", "Vertex Group Input")
+        insertNode(layout, "an_GetVertexGroupNode", "Get Vertex Group")
+        insertNode(layout, "an_InsertVertexGroupNode", "Insert Vertex Group")
         insertNode(layout, "an_SetVertexWeightNode", "Set Vertex Weight")
         insertNode(layout, "an_GetVertexColorLayerNode", "Get Vertex Color Layer")
         insertNode(layout, "an_InsertVertexColorLayerNode", "Insert Vertex Color Layer")